requirements.txt
asyncio
discord.py
expiringdict
iso3166
psycopg2
pyyaml
requests
psutil
